feat(api-client): createApiMerge
